ml/requirements.txt
python-dotenv
groq
langchain
langchain-community
langchainhub
langchain-groq
langchain_openai
langgraph
youtube_search
youtube-transcript-api
pandas
numpy
flask
flask-cors
lxml
imblearn
requests
audio-extract
pytube
pinecone
pinecone-client 
pinecone[grpc]

